add singgle laporan

+++ application/controllers/Buku.php
@@ -14,6 +14,10 @@ class Buku extends CI_Controller
 
     public function add()
     {
+        if (!$this->session->userdata('isLogin') || $this->session->userdata('hak_akses') != 'admin' || $this->session->userdata('hak_akses') != 'keperpus') {
+            redirect(base_url());
+        }
+
         if(isset($_POST)){
             $data = [
                 'kd_buku'=>$this->input->post('kd_buku'),
@@ -51,6 +55,10 @@ class Buku extends CI_Controller
 
     public function update($id)
     {
+        if (!$this->session->userdata('isLogin') || $this->session->userdata('hak_akses') != 'admin' || $this->session->userdata('hak_akses') != 'keperpus') {
+            redirect(base_url());
+        }
+
         if(isset($_POST)){
             $data = [
                 'judul_buku'=>$this->input->post('judul_buku'),
@@ -80,7 +88,7 @@ class Buku extends CI_Controller
                 redirect('/buku');
             }else{
                 $this->session->set_flashdata('message', '<div class="alert alert-danger" role="alert"Maaf, ubah data!</div>');
-            redirect('/buku');
+                redirect('/buku');
             }
         }
     }
@@ -92,6 +100,10 @@ class Buku extends CI_Controller
 
     public function delete($id)
     {
+        if (!$this->session->userdata('isLogin') || $this->session->userdata('hak_akses') != 'admin' || $this->session->userdata('hak_akses') != 'keperpus') {
+            redirect(base_url());
+        }
+        
         if($this->db->delete('buku', $id)){
             $this->session->set_flashdata('message', '<div class="alert alert-success" role="alert">Berhasil Hapus Data!</div>');
             redirect('/buku');
