+++ application/controllers/Laporan.php
@@ -11,6 +11,9 @@ class Laporan extends CI_Controller
 
     public function index()
     {
+        if (!$this->session->userdata('isLogin') || $this->session->userdata('hak_akses') != 'keperpus') {
+            redirect(base_url());
+        }
         $data['title'] = 'Laporan';
         $data['laporan'] = $this->db->order_by('id', 'desc')->get('laporan')->result_array();
         $this->template->load('admin/template', 'admin/laporan/index', $data);
@@ -18,6 +21,9 @@ class Laporan extends CI_Controller
 
     public function buat()
     {
+        if (!$this->session->userdata('isLogin') || $this->session->userdata('hak_akses') != 'keperpus') {
+            redirect(base_url());
+        }
         if(isset($_POST['submit'])){
             $judul_laporan =  $this->input->post('judul_laporan');
             $data = [
@@ -33,7 +39,7 @@ class Laporan extends CI_Controller
             if($tgl_akhir){
                 $data['tgl$tgl_akhir'] = $tgl_akhir;
             }
-            $data2['transaksi'] = $this->m_transaksi->getdata();
+            $data2['transaksi'] = $this->m_transaksi->getlaporan();
             $data2['judul_laporan'] = $judul_laporan;
             $this->load->library('pdf');
             $this->pdf->setPaper('A4', 'landscape');
@@ -49,8 +55,49 @@ class Laporan extends CI_Controller
         }
     }
 
+    public function kembali($id='')
+    {
+        if (!$this->session->userdata('isLogin')) {
+            redirect(base_url());
+        }
+        if($id !== ''){
+            $data['transaksi'] = $this->m_transaksi->getById($id);
+            $data['judul_laporan'] = "Laporan Transaksi - ".$data['transaksi']['nama_siswa'];
+            // var_dump($data); die;
+            $this->load->library('pdf');
+            $this->pdf->setPaper('A4', 'landscape');
+            $this->pdf->filename =$data['judul_laporan'].".pdf";
+            $this->pdf->load_view('admin/laporan/view-singgle', $data);
+        }else{
+            redirect('/');
+        }
+    }
+
+    public function laporanuser($id='')
+    {
+        if (!$this->session->userdata('isLogin')) {
+            redirect(base_url());
+        }
+        if($id !== ''){
+            $no_reg = $this->session->userdata('no_reg');
+            $user = $this->db->get_where('siswa', ['no_reg'=>$no_reg])->row_array();
+            $data['transaksi'] = $this->m_transaksi->getById($id);
+            $data['judul_laporan'] = "Laporan Transaksi - ".$user['nama_siswa'];
+            // var_dump($data); die;
+            $this->load->library('pdf');
+            $this->pdf->setPaper('A4', 'landscape');
+            $this->pdf->filename =$data['judul_laporan'].".pdf";
+            $this->pdf->load_view('admin/laporan/view-singgle', $data);
+        }else{
+            redirect('/');
+        }
+    }
+
     public function delete($id)
     {
+        if (!$this->session->userdata('isLogin') || $this->session->userdata('hak_akses') != 'keperpus') {
+            redirect(base_url());
+        }
         $dokumen = $this->db->get_where('laporan', ['id'=>$id])->row_array();
         $path = 'assets/file/laporan/'.$dokumen['judul_laporan'].'.pdf';
         if($this->db->delete('laporan', ['id'=>$id]) && unlink($path)){
